Switch to SFML

+++ src/PhysicsSystem.h
@@ -1,11 +1,10 @@
 #pragma once
 #include "System.h"
 
-
 class PhysicsSystem : public System {
 public:
-	const double GRAVITY = -9.8;
-	const double SIMSPEED = 1;
+	const float GRAVITY = -9.8f;
+	const float SIMSPEED = 1.0f;
 	void Init();
 
 	void Update(float dt);
